+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Change Log
 ===
 
 
-Version 0.0.6 *(WIP)*
+Version 0.0.6 *(2019-01-21)*
 ---
 
 * Create dedicated object for each operation in a document([#3](https://github.com/yshrsmz/kgql/issues/3))